+++ src/components/Projects.tsx
@@ -8,7 +8,7 @@ import {
   SiSwift,
 } from "react-icons/si";
 import { VscJson } from "react-icons/vsc";
-import ProjectCard from "../sub-components/ProjectCard";
+import ProjectCard from "./sub-components/ProjectCard";
 import { Projects } from "../types/Projects.ts";
 
 const projects: Projects[] = [
